requirements.txt
Django==4.2.3
asgiref==3.6.0
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
django-filter-23.2